WIP: Initial migration to utilize tempest plugin
This is an initial migration step to utilize tempest plugin
system instead of directly calling functional api tests.

This is the approach used by a number of other projects as well
as an approved process by openstack-qa.

The difference in execution is that we will need to execute tempest's
tox instead of our own: tox -eall-plugin api.v1 --  --concurrency=1

I would love feedback on this approach!

Implements: blueprint magnum-tempest
Change-Id: Ic3eadae7fb5d88b776f9ded9589ef25279a2e1be