[ON HOLD] Add magnum certificate api tests
This patch will add cert API test for signing and showing
certificates.

This patch is on hold due to it running over the 2 hour
mark.  It will be unblocked once we lower the time spent
creating bays (right now average time to spin up a bay
ranges between 20 and 35 minutes)

Implements: blueprint magnum-tempest

Change-Id: Ifbb4c779376fa401ca2538aba5097f7af8b4973e